+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased][]
 
+### Added
+
+- Adding error message if target element is not available in the DOM.
+
 ## [9.0.2][] - 2020-02-26
 
 ### Updated
@@ -78,7 +82,5 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 [unreleased]: https://github.com/willmendesneto/ngx-scroll-lock/compare/v9.0.1...HEAD
 [9.0.1]: https://github.com/willmendesneto/ngx-scroll-lock/compare/v9.0.0...v9.0.1
 [9.0.0]: https://github.com/willmendesneto/ngx-scroll-lock/tree/v9.0.0
-
-
-[Unreleased]: https://github.com/willmendesneto/ngx-scroll-lock/compare/v9.0.2...HEAD
+[unreleased]: https://github.com/willmendesneto/ngx-scroll-lock/compare/v9.0.2...HEAD
 [9.0.2]: https://github.com/willmendesneto/ngx-scroll-lock/tree/v9.0.2